auth-service/requirements.txt
fastapi[standard]
uvicorn

sqlalchemy
asyncpg
alembic
psycopg2-binary

python-jose
bcrypt
python-dotenv
pydantic
pydantic-settings

redis
aio-pika